Move event to versioned class
